--- src/Orchard.Web/Modules/Orchard.DynamicForms/ImportExport/FormsExportStep.cs
@@ -1,16 +1,0 @@
-﻿using System.Collections.Generic;
-using Orchard.Environment.Extensions;
-using Orchard.Events;
-
-namespace Orchard.DynamicForms.ImportExport {
-    public interface ICustomExportStep : IEventHandler {
-        void Register(IList<string> steps);
-    }
-
-    [OrchardFeature("Orchard.DynamicForms.ImportExport")]
-    public class FormsExportStep : ICustomExportStep {
-        public void Register(IList<string> steps) {
-            steps.Add("Forms");
-        }
-    }
-}
